♻️ Refactor: move hero into HeroSection component
